Fix issue "SDKException: Connection failure that may be retried."
This issue will happen when your openstacksdk is larger than 0.9.1,
  your cloud service endpoints are using https, and you are using
openstack client to perform actions related to neutron service.

For example :

Command : openstack --debug quota set --floating-ip 2 demo

And you will get errors like this
  "SDKException: Connection failure that may be retried."

This issue is because when performing a HTTP request for Neutron
  service, service endpoint did not get correct, URL schema will be
"http" instead of "https".